contacts objects. An S4 class to represent a file-stored (as .(m)cool) contact matrix imported in R.

contacts S4 class

contacts(
  cool_path,
  resolution = NULL,
  focus = NULL,
  metadata = NULL,
  features = NULL,
  pairs = NULL
)

# S4 method for contacts
length(x)

# S4 method for contacts
dim(x)

# S4 method for contacts,ANY,ANY,ANY
[(x, i)

scores(x, name)

scores(x, name) <- value

features(x, name)

features(x, name) <- value

pairsFile(x, name)

pairsFile(x) <- value

# S4 method for contacts
anchors(x)

# S4 method for contacts
regions(x)

.summary(object)

# S4 method for contacts
show(object)

Arguments

cool_path

Path of a (m)cool file

resolution

Resolution to use with mcool file

focus

focus Chr. coordinates for which interaction counts are extracted from the .(m)cool file. Can be provided as a character string or as a GRanges object

metadata

list of metadata

features

features provided as a named SimpleList

pairs

Path to an associated .pairs file

x

A contacts object.

i

a range or boolean vector.

name

name

value

value

object

A contacts object.

Value

a new contacts object.

Slots

focus

Chr. coordinates for which interaction counts are extracted from the .(m)cool file.

metadata

<etadata associated with the .(m)cool file.

seqinfo

Seqinfo, deduced from the largest resolution available in the .(m)cool file.

resolutions

Resolutions available in the .(m)cool file.

current_resolution

Current resolution

bins

Genomic bins (computed on-the-fly) of seqinfo at the current resolution

interactions

Genomic Interactions extracted from the .(m)cool object

scores

Available interaction scores.

features

Genomic features associated with the dataset (e.g. loops, borders, etc...)

pairsFile

path for the .pairs file associated with the .(m)cool file

type

Optional. Type of contacts matrix (sparse, full, aggr, ratio, ...)

Examples

library(HiContacts)
data(contacts_yeast)
contacts_yeast
#> sparse `contacts` object with 74,360 interactions over 802 regions 
#> file: /home/rsg/.cache/R/ExperimentHub/yeast_wt.mcool 
#> focus: II 
#> ------------
#> metadata(1): path
#> resolutions(5): 1000 2000 4000 8000 16000
#> current resolution(1000): 74360 interactions in memory 
#> bins: 12079 
#> regions: 802 
#> scores(2): raw balanced 
#> features: loops(0) borders(0) compartments(0) viewpoints(0) 
#> pairs: N/A 
library(HiContacts)
data(contacts_yeast)
length(contacts_yeast)
#> [1] 74360
library(HiContacts)
data(contacts_yeast)
dim(contacts_yeast)
#> [1] 802 802
library(HiContacts)
data(contacts_yeast)
contacts_yeast[1:10]
#> sparse `contacts` object with 10 interactions over 802 regions 
#> file: /home/rsg/.cache/R/ExperimentHub/yeast_wt.mcool 
#> focus: II 
#> ------------
#> metadata(1): path
#> resolutions(5): 1000 2000 4000 8000 16000
#> current resolution(1000): 10 interactions in memory 
#> bins: 12079 
#> regions: 802 
#> scores(2): raw balanced 
#> features: loops(0) borders(0) compartments(0) viewpoints(0) 
#> pairs: N/A 
library(HiContacts)
data(contacts_yeast)
scores(contacts_yeast)
#> List of length 2
#> names(2): raw balanced
library(HiContacts)
data(contacts_yeast)
scores(contacts_yeast, 'test') <- runif(length(contacts_yeast))
scores(contacts_yeast, 'test')
#> GInteractions object with 74360 interactions and 1 metadata column:
#>           seqnames1       ranges1     seqnames2       ranges2 |     score
#>               <Rle>     <IRanges>         <Rle>     <IRanges> | <numeric>
#>       [1]        II        1-1000 ---        II     1001-2000 | 0.0913248
#>       [2]        II        1-1000 ---        II     5001-6000 | 0.8633831
#>       [3]        II        1-1000 ---        II     6001-7000 | 0.4485691
#>       [4]        II        1-1000 ---        II     8001-9000 | 0.5450900
#>       [5]        II        1-1000 ---        II    9001-10000 | 0.3067725
#>       ...       ...           ... ...       ...           ... .       ...
#>   [74356]        II 807001-808000 ---        II 809001-810000 |  0.939865
#>   [74357]        II 807001-808000 ---        II 810001-811000 |  0.944560
#>   [74358]        II 808001-809000 ---        II 808001-809000 |  0.100953
#>   [74359]        II 808001-809000 ---        II 809001-810000 |  0.486821
#>   [74360]        II 809001-810000 ---        II 809001-810000 |  0.432076
#>   -------
#>   regions: 802 ranges and 2 metadata columns
#>   seqinfo: 16 sequences from an unspecified genome
library(HiContacts)
data(contacts_yeast)
features(contacts_yeast)
#> List of length 4
#> names(4): loops borders compartments viewpoints
library(HiContacts)
data(contacts_yeast)
data(centros_yeast)
features(contacts_yeast, 'centromeres') <- centros_yeast
features(contacts_yeast, 'centromeres')
#> GRanges object with 16 ranges and 0 metadata columns:
#>        seqnames        ranges strand
#>           <Rle>     <IRanges>  <Rle>
#>    [1]        I 151583-151641      +
#>    [2]       II 238361-238419      +
#>    [3]      III 114322-114380      +
#>    [4]       IV 449879-449937      +
#>    [5]        V 152522-152580      +
#>    ...      ...           ...    ...
#>   [12]      XII 151366-151424      +
#>   [13]     XIII 268222-268280      +
#>   [14]      XIV 628588-628646      +
#>   [15]       XV 326897-326955      +
#>   [16]      XVI 556255-556313      +
#>   -------
#>   seqinfo: 17 sequences (1 circular) from R64-1-1 genome
library(HiContacts)
data(full_contacts_yeast)
pairsFile(full_contacts_yeast)
#> NULL
library(HiContacts)
data(full_contacts_yeast)
anchors(full_contacts_yeast)
#> $first
#> GRanges object with 267702 ranges and 2 metadata columns:
#>            seqnames        ranges strand |   chr    center
#>               <Rle>     <IRanges>  <Rle> | <Rle> <integer>
#>        [1]        I       1-16000      * |     I      8000
#>        [2]        I       1-16000      * |     I      8000
#>        [3]        I       1-16000      * |     I      8000
#>        [4]        I       1-16000      * |     I      8000
#>        [5]        I       1-16000      * |     I      8000
#>        ...      ...           ...    ... .   ...       ...
#>   [267698]      XVI 896001-912000      * |   XVI    904000
#>   [267699]      XVI 896001-912000      * |   XVI    904000
#>   [267700]      XVI 912001-928000      * |   XVI    920000
#>   [267701]      XVI 912001-928000      * |   XVI    920000
#>   [267702]      XVI 928001-944000      * |   XVI    936000
#>   -------
#>   seqinfo: 16 sequences from an unspecified genome
#> 
#> $second
#> GRanges object with 267702 ranges and 2 metadata columns:
#>            seqnames        ranges strand |   chr    center
#>               <Rle>     <IRanges>  <Rle> | <Rle> <integer>
#>        [1]        I       1-16000      * |     I      8000
#>        [2]        I   16001-32000      * |     I     24000
#>        [3]        I   32001-48000      * |     I     40000
#>        [4]        I   48001-64000      * |     I     56000
#>        [5]        I   64001-80000      * |     I     72000
#>        ...      ...           ...    ... .   ...       ...
#>   [267698]      XVI 912001-928000      * |   XVI    920000
#>   [267699]      XVI 928001-944000      * |   XVI    936000
#>   [267700]      XVI 912001-928000      * |   XVI    920000
#>   [267701]      XVI 928001-944000      * |   XVI    936000
#>   [267702]      XVI 928001-944000      * |   XVI    936000
#>   -------
#>   seqinfo: 16 sequences from an unspecified genome
#> 
library(HiContacts)
data(full_contacts_yeast)
regions(full_contacts_yeast)
#> GRanges object with 757 ranges and 2 metadata columns:
#>                     seqnames        ranges strand |   chr    center
#>                        <Rle>     <IRanges>  <Rle> | <Rle> <integer>
#>           I_1_16000        I       1-16000      * |     I      8000
#>       I_16001_32000        I   16001-32000      * |     I     24000
#>       I_32001_48000        I   32001-48000      * |     I     40000
#>       I_48001_64000        I   48001-64000      * |     I     56000
#>       I_64001_80000        I   64001-80000      * |     I     72000
#>                 ...      ...           ...    ... .   ...       ...
#>   XVI_864001_880000      XVI 864001-880000      * |   XVI    872000
#>   XVI_880001_896000      XVI 880001-896000      * |   XVI    888000
#>   XVI_896001_912000      XVI 896001-912000      * |   XVI    904000
#>   XVI_912001_928000      XVI 912001-928000      * |   XVI    920000
#>   XVI_928001_944000      XVI 928001-944000      * |   XVI    936000
#>   -------
#>   seqinfo: 16 sequences from an unspecified genome
library(HiContacts)
data(contacts_yeast)
summary(contacts_yeast)
#>   Length    Class     Mode 
#>    74360 contacts       S4 
library(HiContacts)
data(contacts_yeast)
contacts_yeast
#> sparse `contacts` object with 74,360 interactions over 802 regions 
#> file: /home/rsg/.cache/R/ExperimentHub/yeast_wt.mcool 
#> focus: II 
#> ------------
#> metadata(1): path
#> resolutions(5): 1000 2000 4000 8000 16000
#> current resolution(1000): 74360 interactions in memory 
#> bins: 12079 
#> regions: 802 
#> scores(3): raw balanced test 
#> features: loops(0) borders(0) compartments(0) viewpoints(0) centromeres(16) 
#> pairs: N/A