Compute the law of distance-dependent contact frequency, a.k.a. P(s)

P(s) will be approximated if no pairs are provided, or the exact P(s) will be computed if a .pairs file is added to the Contacts object using pairsFile(x) <- "...".

distanceLaw(
  x,
  by_chr = FALSE,
  filtered_chr = c("XII", "chrXII", "chr12", "12", "Mito", "MT", "chrM")
)

PsBreaks()

Arguments

x

A Contacts object

by_chr

by_chr

filtered_chr

filtered_chr

Value

a tibble

tbl

Examples

library(HiContacts)
contacts_yeast <- contacts_yeast()
#> snapshotDate(): 2022-08-23
#> see ?HiContactsData and browseVignettes('HiContactsData') for documentation
#> loading from cache
ps <- distanceLaw(contacts_yeast)
#> pairsFile not specified. The P(s) curve will be an approximation.
ps
#> # A tibble: 57 × 5
#>    binned_distance      p    norm_p norm_p_unity slope
#>              <dbl>  <dbl>     <dbl>        <dbl> <dbl>
#>  1               1 0.121  0.121         242583.   0   
#>  2             956 0.0735 0.000774        1557.  -1.68
#>  3            1862 0.0700 0.000376         757.  -1.63
#>  4            2999 0.0576 0.000192         386.  -1.59
#>  5            3992 0.0462 0.000116         233.  -1.53
#>  6            4830 0.0386 0.0000800        161.  -1.48
#>  7            5844 0.0336 0.0000575        116.  -1.44
#>  8            6429 0.0292 0.0000453         91.2 -1.40
#>  9            7779 0.0263 0.0000339         68.1 -1.37
#> 10            8557 0.0240 0.0000281         56.5 -1.32
#> # … with 47 more rows