contacts S4 class and methods

contacts S4 class and methods

contacts(
  file,
  resolution = NULL,
  focus = NULL,
  metadata = list(),
  topologicalFeatures = S4Vectors::SimpleList(loops =
    S4Vectors::Pairs(GenomicRanges::GRanges(), GenomicRanges::GRanges()), borders =
    GenomicRanges::GRanges(), compartments = GenomicRanges::GRanges(), viewpoints =
    GenomicRanges::GRanges()),
  pairsFile = NULL
)

# S4 method for contacts
length(x)

# S4 method for contacts,numeric,ANY,ANY
[(x, i)

# S4 method for contacts
fileName(object)

# S4 method for contacts
seqinfo(x)

resolutions(x)

# S4 method for contacts
resolution(x)

bins(x)

focus(x)

# S4 method for contacts
interactions(x)

# S4 method for contacts,GInteractions
interactions(x) <- value

scores(x, name)

scores(x, name) <- value

topologicalFeatures(x, name)

topologicalFeatures(x, name) <- value

pairsFile(x, name)

pairsFile(x) <- value

metadata(x) <- value

# S4 method for contacts
anchors(x)

# S4 method for contacts
regions(x)

# S4 method for contacts
summary(object)

# S4 method for contacts
show(object)

Arguments

file

Path to a (m)cool file

resolution

Resolution to use with mcool file

focus

focus Chr. coordinates for which interaction counts are extracted from the .(m)cool file, provided as a character string (e.g. "II:4000-5000"). If not provided, the entire (m)cool file will be imported.

metadata

list of metadata

topologicalFeatures

topologicalFeatures provided as a named SimpleList

pairsFile

Path to an associated .pairs file

x

A contacts object.

i

a range or boolean vector.

object

A contacts object.

value

value

name

name

Value

a new contacts object.

Slots

fileName

Path of (m)cool file

focus

Chr. coordinates for which interaction counts are extracted from the .(m)cool file.

resolutions

Resolutions available in the .(m)cool file.

resolution

Current resolution

interactions

Genomic Interactions extracted from the .(m)cool object

scores

Available interaction scores.

topologicalFeatures

Topological features associated with the dataset (e.g. loops (\<Pairs\>), borders (\<GRanges\>), viewpoints (\<GRanges\>), etc...)

pairsFile

Path to the .pairs file associated with the .(m)cool file

metadata

metadata associated with the .(m)cool file.

Examples

library(HiContacts)
contacts_yeast <- contacts_yeast()
#> snapshotDate(): 2022-08-23
#> see ?HiContactsData and browseVignettes('HiContactsData') for documentation
#> loading from cache
contacts_yeast
#> `contacts` object with 74,360 interactions over 802 regions 
#> -------
#> fileName: "/github/home/.cache/R/ExperimentHub/3667eb5d688_7752" 
#> focus: "II" 
#> resolutions(5): 1000 2000 4000 8000 16000
#> current resolution: 1000 
#> interactions: 74360 
#> scores(2): raw balanced 
#> topologicalFeatures: loops(0) borders(0) compartments(0) viewpoints(0) 
#> pairsFile: N/A 
#> metadata(0):
length(contacts_yeast)
#> [1] 74360
contacts_yeast[seq_len(10)]
#> `contacts` object with 10 interactions over 802 regions 
#> -------
#> fileName: "/github/home/.cache/R/ExperimentHub/3667eb5d688_7752" 
#> focus: "II" 
#> resolutions(5): 1000 2000 4000 8000 16000
#> current resolution: 1000 
#> interactions: 10 
#> scores(2): raw balanced 
#> topologicalFeatures: loops(0) borders(0) compartments(0) viewpoints(0) 
#> pairsFile: N/A 
#> metadata(0):
fileName(contacts_yeast)
#> [1] "/github/home/.cache/R/ExperimentHub/3667eb5d688_7752"
seqinfo(contacts_yeast)
#> Seqinfo object with 16 sequences from an unspecified genome:
#>   seqnames seqlengths isCircular genome
#>   I            230218       <NA>   <NA>
#>   II           813184       <NA>   <NA>
#>   III          316620       <NA>   <NA>
#>   IV          1531933       <NA>   <NA>
#>   V            576874       <NA>   <NA>
#>   ...             ...        ...    ...
#>   XII         1078177       <NA>   <NA>
#>   XIII         924431       <NA>   <NA>
#>   XIV          784333       <NA>   <NA>
#>   XV          1091291       <NA>   <NA>
#>   XVI          948066       <NA>   <NA>
resolutions(contacts_yeast)
#> [1]  1000  2000  4000  8000 16000
resolution(contacts_yeast)
#> [1] 1000
bins(contacts_yeast)
#> GRanges object with 12079 ranges and 1 metadata column:
#>           seqnames        ranges strand |    bin_id
#>              <Rle>     <IRanges>  <Rle> | <integer>
#>       [1]        I        1-1000      * |         1
#>       [2]        I     1001-2000      * |         2
#>       [3]        I     2001-3000      * |         3
#>       [4]        I     3001-4000      * |         4
#>       [5]        I     4001-5000      * |         5
#>       ...      ...           ...    ... .       ...
#>   [12075]      XVI 944001-945000      * |     12075
#>   [12076]      XVI 945001-946000      * |     12076
#>   [12077]      XVI 946001-947000      * |     12077
#>   [12078]      XVI 947001-948000      * |     12078
#>   [12079]      XVI 948001-948066      * |     12079
#>   -------
#>   seqinfo: 16 sequences from an unspecified genome
focus(contacts_yeast)
#> [1] "II"
interactions(contacts_yeast)
#> GInteractions object with 74360 interactions and 0 metadata columns:
#>           seqnames1       ranges1     seqnames2       ranges2
#>               <Rle>     <IRanges>         <Rle>     <IRanges>
#>       [1]        II        1-1000 ---        II     1001-2000
#>       [2]        II        1-1000 ---        II     5001-6000
#>       [3]        II        1-1000 ---        II     6001-7000
#>       [4]        II        1-1000 ---        II     8001-9000
#>       [5]        II        1-1000 ---        II    9001-10000
#>       ...       ...           ... ...       ...           ...
#>   [74356]        II 807001-808000 ---        II 809001-810000
#>   [74357]        II 807001-808000 ---        II 810001-811000
#>   [74358]        II 808001-809000 ---        II 808001-809000
#>   [74359]        II 808001-809000 ---        II 809001-810000
#>   [74360]        II 809001-810000 ---        II 809001-810000
#>   -------
#>   regions: 802 ranges and 2 metadata columns
#>   seqinfo: 16 sequences from an unspecified genome
scores(contacts_yeast)
#> List of length 2
#> names(2): raw balanced
tail(scores(contacts_yeast, 1))
#> [1] 14  8  1  1  7  2
tail(scores(contacts_yeast, 'balanced'))
#> [1]        NaN 0.04720230        NaN        NaN        NaN 0.04113549
scores(contacts_yeast, 'test') <- runif(length(contacts_yeast))
tail(scores(contacts_yeast, 'test'))
#> [1] 0.4340942 0.0366653 0.7579076 0.1420236 0.2873017 0.4189317
full_contacts_yeast <- full_contacts_yeast()
#> snapshotDate(): 2022-08-23
#> see ?HiContactsData and browseVignettes('HiContactsData') for documentation
#> loading from cache
topologicalFeatures(full_contacts_yeast)
#> List of length 5
#> names(5): loops borders compartments viewpoints centromeres
topologicalFeatures(full_contacts_yeast, 1)
#> Pairs object with 0 pairs and 0 metadata columns:
#>        first    second
#>    <GRanges> <GRanges>
topologicalFeatures(full_contacts_yeast, 'centromeres')
#> GRanges object with 16 ranges and 0 metadata columns:
#>        seqnames        ranges strand
#>           <Rle>     <IRanges>  <Rle>
#>    [1]        I 151583-151641      +
#>    [2]       II 238361-238419      +
#>    [3]      III 114322-114380      +
#>    [4]       IV 449879-449937      +
#>    [5]        V 152522-152580      +
#>    ...      ...           ...    ...
#>   [12]      XII 151366-151424      +
#>   [13]     XIII 268222-268280      +
#>   [14]      XIV 628588-628646      +
#>   [15]       XV 326897-326955      +
#>   [16]      XVI 556255-556313      +
#>   -------
#>   seqinfo: 17 sequences (1 circular) from R64-1-1 genome
data(centros_yeast)
topologicalFeatures(contacts_yeast, 'centromeres') <- centros_yeast
topologicalFeatures(contacts_yeast, 'centromeres')
#> GRanges object with 16 ranges and 0 metadata columns:
#>        seqnames        ranges strand
#>           <Rle>     <IRanges>  <Rle>
#>    [1]        I 151583-151641      +
#>    [2]       II 238361-238419      +
#>    [3]      III 114322-114380      +
#>    [4]       IV 449879-449937      +
#>    [5]        V 152522-152580      +
#>    ...      ...           ...    ...
#>   [12]      XII 151366-151424      +
#>   [13]     XIII 268222-268280      +
#>   [14]      XIV 628588-628646      +
#>   [15]       XV 326897-326955      +
#>   [16]      XVI 556255-556313      +
#>   -------
#>   seqinfo: 17 sequences (1 circular) from R64-1-1 genome
pairsFile(full_contacts_yeast)
#> NULL
anchors(contacts_yeast)
#> $first
#> GRanges object with 74360 ranges and 2 metadata columns:
#>           seqnames        ranges strand |   chr    center
#>              <Rle>     <IRanges>  <Rle> | <Rle> <integer>
#>       [1]       II        1-1000      * |    II       500
#>       [2]       II        1-1000      * |    II       500
#>       [3]       II        1-1000      * |    II       500
#>       [4]       II        1-1000      * |    II       500
#>       [5]       II        1-1000      * |    II       500
#>       ...      ...           ...    ... .   ...       ...
#>   [74356]       II 807001-808000      * |    II    807500
#>   [74357]       II 807001-808000      * |    II    807500
#>   [74358]       II 808001-809000      * |    II    808500
#>   [74359]       II 808001-809000      * |    II    808500
#>   [74360]       II 809001-810000      * |    II    809500
#>   -------
#>   seqinfo: 16 sequences from an unspecified genome
#> 
#> $second
#> GRanges object with 74360 ranges and 2 metadata columns:
#>           seqnames        ranges strand |   chr    center
#>              <Rle>     <IRanges>  <Rle> | <Rle> <integer>
#>       [1]       II     1001-2000      * |    II      1500
#>       [2]       II     5001-6000      * |    II      5500
#>       [3]       II     6001-7000      * |    II      6500
#>       [4]       II     8001-9000      * |    II      8500
#>       [5]       II    9001-10000      * |    II      9500
#>       ...      ...           ...    ... .   ...       ...
#>   [74356]       II 809001-810000      * |    II    809500
#>   [74357]       II 810001-811000      * |    II    810500
#>   [74358]       II 808001-809000      * |    II    808500
#>   [74359]       II 809001-810000      * |    II    809500
#>   [74360]       II 809001-810000      * |    II    809500
#>   -------
#>   seqinfo: 16 sequences from an unspecified genome
#> 
regions(contacts_yeast)
#> GRanges object with 802 ranges and 2 metadata columns:
#>                    seqnames        ranges strand |   chr    center
#>                       <Rle>     <IRanges>  <Rle> | <Rle> <integer>
#>          II_1_1000       II        1-1000      * |    II       500
#>       II_1001_2000       II     1001-2000      * |    II      1500
#>       II_5001_6000       II     5001-6000      * |    II      5500
#>       II_6001_7000       II     6001-7000      * |    II      6500
#>       II_7001_8000       II     7001-8000      * |    II      7500
#>                ...      ...           ...    ... .   ...       ...
#>   II_807001_808000       II 807001-808000      * |    II    807500
#>   II_808001_809000       II 808001-809000      * |    II    808500
#>   II_809001_810000       II 809001-810000      * |    II    809500
#>   II_810001_811000       II 810001-811000      * |    II    810500
#>   II_812001_813000       II 812001-813000      * |    II    812500
#>   -------
#>   seqinfo: 16 sequences from an unspecified genome
summary(contacts_yeast)
#> `contacts` object with 74,360 interactions over 802 regions 
show(contacts_yeast)
#> `contacts` object with 74,360 interactions over 802 regions 
#> -------
#> fileName: "/github/home/.cache/R/ExperimentHub/3667eb5d688_7752" 
#> focus: "II" 
#> resolutions(5): 1000 2000 4000 8000 16000
#> current resolution: 1000 
#> interactions: 74360 
#> scores(3): raw balanced test 
#> topologicalFeatures: loops(0) borders(0) compartments(0) viewpoints(0) centromeres(16) 
#> pairsFile: N/A 
#> metadata(0):
as(contacts_yeast, 'GInteractions')
#> GInteractions object with 74360 interactions and 0 metadata columns:
#>           seqnames1       ranges1     seqnames2       ranges2
#>               <Rle>     <IRanges>         <Rle>     <IRanges>
#>       [1]        II        1-1000 ---        II     1001-2000
#>       [2]        II        1-1000 ---        II     5001-6000
#>       [3]        II        1-1000 ---        II     6001-7000
#>       [4]        II        1-1000 ---        II     8001-9000
#>       [5]        II        1-1000 ---        II    9001-10000
#>       ...       ...           ... ...       ...           ...
#>   [74356]        II 807001-808000 ---        II 809001-810000
#>   [74357]        II 807001-808000 ---        II 810001-811000
#>   [74358]        II 808001-809000 ---        II 808001-809000
#>   [74359]        II 808001-809000 ---        II 809001-810000
#>   [74360]        II 809001-810000 ---        II 809001-810000
#>   -------
#>   regions: 802 ranges and 2 metadata columns
#>   seqinfo: 16 sequences from an unspecified genome
as(contacts_yeast, 'ContactMatrix')
#> class: ContactMatrix 
#> dim: 802 802 
#> type: matrix 
#> rownames: NULL
#> colnames: NULL
#> metadata(0):
#> regions: 802
as(contacts_yeast, 'matrix')[seq_len(10), seq_len(10)]
#>       [,1] [,2]       [,3] [,4] [,5]       [,6]       [,7]       [,8]
#>  [1,]   NA   NA         NA   NA   NA         NA         NA         NA
#>  [2,]   NA   NA         NA   NA   NA         NA         NA         NA
#>  [3,]   NA   NA         NA   NA   NA 0.02617935 0.08584875 0.06135093
#>  [4,]   NA   NA         NA   NA   NA         NA         NA         NA
#>  [5,]   NA   NA         NA   NA   NA         NA         NA         NA
#>  [6,]   NA   NA 0.02617935   NA   NA 0.05122552 0.05990950 0.09589496
#>  [7,]   NA   NA 0.08584875   NA   NA 0.05990950 0.02521393 0.05526944
#>  [8,]   NA   NA 0.06135093   NA   NA 0.09589496 0.05526944 0.01959171
#>  [9,]   NA   NA 0.02450783   NA   NA 0.05256588 0.06928067 0.06583294
#> [10,]   NA   NA 0.04734105   NA   NA 0.05344208 0.05735461 0.05394987
#>             [,9]      [,10]
#>  [1,]         NA         NA
#>  [2,]         NA         NA
#>  [3,] 0.02450783 0.04734105
#>  [4,]         NA         NA
#>  [5,]         NA         NA
#>  [6,] 0.05256588 0.05344208
#>  [7,] 0.06928067 0.05735461
#>  [8,] 0.06583294 0.05394987
#>  [9,] 0.02330981 0.08755227
#> [10,] 0.08755227 0.03221374