contacts S4 class and methods

contacts S4 class and methods

features method for objects of class contacts.

features<- method for objects of class contacts.

pairsFile method for objects of class contacts.

pairsFile<- method for objects of class contacts.

anchors method for objects of class contacts.

regions method for objects of class contacts.

summary method for objects of class contacts.

Show method for objects of class contacts.

contacts(
  cool_path,
  resolution = NULL,
  focus = NULL,
  metadata = list(),
  features = S4Vectors::SimpleList(loops = GenomicRanges::GRanges(), borders =
    GenomicRanges::GRanges(), compartments = GenomicRanges::GRanges(), viewpoints =
    GenomicRanges::GRanges()),
  pairs = NULL
)

# S4 method for contacts
length(x)

# S4 method for contacts,ANY,ANY,ANY
[(x, i)

matrixType(x)

coolPath(x)

# S4 method for contacts
seqinfo(x)

resolutions(x)

# S4 method for contacts
resolution(x)

bins(x)

focus(x)

# S4 method for contacts
interactions(x)

scores(x, name)

scores(x, name) <- value

features(x, name)

features(x, name) <- value

pairsFile(x, name)

pairsFile(x) <- value

# S4 method for contacts
anchors(x)

# S4 method for contacts
regions(x)

# S4 method for contacts
summary(object)

# S4 method for contacts
show(object)

Arguments

cool_path

Path of a (m)cool file

resolution

Resolution to use with mcool file

focus

focus Chr. coordinates for which interaction counts are extracted from the .(m)cool file. Can be provided as a character string or as a GRanges object

metadata

list of metadata

features

features provided as a named SimpleList

pairs

Path to an associated .pairs file

x

A contacts object.

i

a range or boolean vector.

name

name

value

value

object

A contacts object.

Value

a new contacts object.

Slots

focus

Chr. coordinates for which interaction counts are extracted from the .(m)cool file.

metadata

metadata associated with the .(m)cool file.

seqinfo

Seqinfo, deduced from the largest resolution available in the .(m)cool file.

resolutions

Resolutions available in the .(m)cool file.

current_resolution

Current resolution

bins

Genomic bins (computed on-the-fly) of seqinfo at the current resolution

interactions

Genomic Interactions extracted from the .(m)cool object

scores

Available interaction scores.

features

Genomic features associated with the dataset (e.g. loops, borders, etc...)

pairsFile

path for the .pairs file associated with the .(m)cool file

matrixType

Type of contacts matrix (sparse, full, aggr, ratio, ...)

coolPath

Path of (m)cool file

Examples

library(HiContacts)
data(contacts_yeast)
contacts_yeast
#> sparse `contacts` object with 74,360 interactions over 802 regions 
#> coolPath: /home/rsg/.cache/R/ExperimentHub/36d590c5583_7752 
#> focus: II 
#> ------------
#> metadata(0):
#> resolutions(5): 1000 2000 4000 8000 16000
#> current resolution(1000): 74360 interactions in memory 
#> bins: 12079 
#> regions: 802 
#> scores(2): raw balanced 
#> features: loops(0) borders(0) compartments(0) viewpoints(0) 
#> pairs: N/A 
length(contacts_yeast)
#> [1] 74360
contacts_yeast[1:10]
#> sparse `contacts` object with 10 interactions over 802 regions 
#> coolPath: /home/rsg/.cache/R/ExperimentHub/36d590c5583_7752 
#> focus: II 
#> ------------
#> metadata(0):
#> resolutions(5): 1000 2000 4000 8000 16000
#> current resolution(1000): 10 interactions in memory 
#> bins: 12079 
#> regions: 802 
#> scores(2): raw balanced 
#> features: loops(0) borders(0) compartments(0) viewpoints(0) 
#> pairs: N/A 
matrixType(contacts_yeast)
#> [1] "sparse"
coolPath(contacts_yeast)
#> [1] "/home/rsg/.cache/R/ExperimentHub/36d590c5583_7752"
seqinfo(contacts_yeast)
#> Seqinfo object with 16 sequences from an unspecified genome:
#>   seqnames seqlengths isCircular genome
#>   I            230218       <NA>   <NA>
#>   II           813184       <NA>   <NA>
#>   III          316620       <NA>   <NA>
#>   IV          1531933       <NA>   <NA>
#>   V            576874       <NA>   <NA>
#>   ...             ...        ...    ...
#>   XII         1078177       <NA>   <NA>
#>   XIII         924431       <NA>   <NA>
#>   XIV          784333       <NA>   <NA>
#>   XV          1091291       <NA>   <NA>
#>   XVI          948066       <NA>   <NA>
resolutions(contacts_yeast)
#> [1]  1000  2000  4000  8000 16000
resolution(contacts_yeast)
#> [1] 1000
bins(contacts_yeast)
#> GRanges object with 12079 ranges and 0 metadata columns:
#>           seqnames        ranges strand
#>              <Rle>     <IRanges>  <Rle>
#>       [1]        I        1-1000      *
#>       [2]        I     1001-2000      *
#>       [3]        I     2001-3000      *
#>       [4]        I     3001-4000      *
#>       [5]        I     4001-5000      *
#>       ...      ...           ...    ...
#>   [12075]      XVI 944001-945000      *
#>   [12076]      XVI 945001-946000      *
#>   [12077]      XVI 946001-947000      *
#>   [12078]      XVI 947001-948000      *
#>   [12079]      XVI 948001-948066      *
#>   -------
#>   seqinfo: 16 sequences from an unspecified genome
focus(contacts_yeast)
#> [1] "II"
interactions(contacts_yeast)
#> GInteractions object with 74360 interactions and 0 metadata columns:
#>           seqnames1       ranges1     seqnames2       ranges2
#>               <Rle>     <IRanges>         <Rle>     <IRanges>
#>       [1]        II        1-1000 ---        II     1001-2000
#>       [2]        II        1-1000 ---        II     5001-6000
#>       [3]        II        1-1000 ---        II     6001-7000
#>       [4]        II        1-1000 ---        II     8001-9000
#>       [5]        II        1-1000 ---        II    9001-10000
#>       ...       ...           ... ...       ...           ...
#>   [74356]        II 807001-808000 ---        II 809001-810000
#>   [74357]        II 807001-808000 ---        II 810001-811000
#>   [74358]        II 808001-809000 ---        II 808001-809000
#>   [74359]        II 808001-809000 ---        II 809001-810000
#>   [74360]        II 809001-810000 ---        II 809001-810000
#>   -------
#>   regions: 802 ranges and 2 metadata columns
#>   seqinfo: 16 sequences from an unspecified genome
scores(contacts_yeast)
#> List of length 2
#> names(2): raw balanced
scores(contacts_yeast, 1)
#> GInteractions object with 74360 interactions and 1 metadata column:
#>           seqnames1       ranges1     seqnames2       ranges2 |     score
#>               <Rle>     <IRanges>         <Rle>     <IRanges> | <integer>
#>       [1]        II        1-1000 ---        II     1001-2000 |         1
#>       [2]        II        1-1000 ---        II     5001-6000 |         2
#>       [3]        II        1-1000 ---        II     6001-7000 |         1
#>       [4]        II        1-1000 ---        II     8001-9000 |         2
#>       [5]        II        1-1000 ---        II    9001-10000 |         3
#>       ...       ...           ... ...       ...           ... .       ...
#>   [74356]        II 807001-808000 ---        II 809001-810000 |         8
#>   [74357]        II 807001-808000 ---        II 810001-811000 |         1
#>   [74358]        II 808001-809000 ---        II 808001-809000 |         1
#>   [74359]        II 808001-809000 ---        II 809001-810000 |         7
#>   [74360]        II 809001-810000 ---        II 809001-810000 |         2
#>   -------
#>   regions: 802 ranges and 2 metadata columns
#>   seqinfo: 16 sequences from an unspecified genome
scores(contacts_yeast, 'balanced')
#> GInteractions object with 74360 interactions and 1 metadata column:
#>           seqnames1       ranges1     seqnames2       ranges2 |     score
#>               <Rle>     <IRanges>         <Rle>     <IRanges> | <numeric>
#>       [1]        II        1-1000 ---        II     1001-2000 |       NaN
#>       [2]        II        1-1000 ---        II     5001-6000 |       NaN
#>       [3]        II        1-1000 ---        II     6001-7000 |       NaN
#>       [4]        II        1-1000 ---        II     8001-9000 |       NaN
#>       [5]        II        1-1000 ---        II    9001-10000 |       NaN
#>       ...       ...           ... ...       ...           ... .       ...
#>   [74356]        II 807001-808000 ---        II 809001-810000 | 0.0472023
#>   [74357]        II 807001-808000 ---        II 810001-811000 |       NaN
#>   [74358]        II 808001-809000 ---        II 808001-809000 |       NaN
#>   [74359]        II 808001-809000 ---        II 809001-810000 |       NaN
#>   [74360]        II 809001-810000 ---        II 809001-810000 | 0.0411355
#>   -------
#>   regions: 802 ranges and 2 metadata columns
#>   seqinfo: 16 sequences from an unspecified genome
scores(contacts_yeast, 'test') <- runif(length(contacts_yeast))
scores(contacts_yeast, 'test')
#> GInteractions object with 74360 interactions and 1 metadata column:
#>           seqnames1       ranges1     seqnames2       ranges2 |     score
#>               <Rle>     <IRanges>         <Rle>     <IRanges> | <numeric>
#>       [1]        II        1-1000 ---        II     1001-2000 |  0.781219
#>       [2]        II        1-1000 ---        II     5001-6000 |  0.483858
#>       [3]        II        1-1000 ---        II     6001-7000 |  0.710096
#>       [4]        II        1-1000 ---        II     8001-9000 |  0.298953
#>       [5]        II        1-1000 ---        II    9001-10000 |  0.266175
#>       ...       ...           ... ...       ...           ... .       ...
#>   [74356]        II 807001-808000 ---        II 809001-810000 |  0.884464
#>   [74357]        II 807001-808000 ---        II 810001-811000 |  0.707703
#>   [74358]        II 808001-809000 ---        II 808001-809000 |  0.730158
#>   [74359]        II 808001-809000 ---        II 809001-810000 |  0.282583
#>   [74360]        II 809001-810000 ---        II 809001-810000 |  0.454623
#>   -------
#>   regions: 802 ranges and 2 metadata columns
#>   seqinfo: 16 sequences from an unspecified genome
data(full_contacts_yeast)
features(full_contacts_yeast)
#> List of length 5
#> names(5): loops borders compartments viewpoints centromeres
features(full_contacts_yeast, 1)
#> GRanges object with 0 ranges and 0 metadata columns:
#>    seqnames    ranges strand
#>       <Rle> <IRanges>  <Rle>
#>   -------
#>   seqinfo: no sequences
features(full_contacts_yeast, 'centromeres')
#> GRanges object with 16 ranges and 0 metadata columns:
#>        seqnames        ranges strand
#>           <Rle>     <IRanges>  <Rle>
#>    [1]        I 151583-151641      +
#>    [2]       II 238361-238419      +
#>    [3]      III 114322-114380      +
#>    [4]       IV 449879-449937      +
#>    [5]        V 152522-152580      +
#>    ...      ...           ...    ...
#>   [12]      XII 151366-151424      +
#>   [13]     XIII 268222-268280      +
#>   [14]      XIV 628588-628646      +
#>   [15]       XV 326897-326955      +
#>   [16]      XVI 556255-556313      +
#>   -------
#>   seqinfo: 17 sequences (1 circular) from R64-1-1 genome
data(centros_yeast)
features(contacts_yeast, 'centromeres') <- centros_yeast
features(contacts_yeast, 'centromeres')
#> GRanges object with 16 ranges and 0 metadata columns:
#>        seqnames        ranges strand
#>           <Rle>     <IRanges>  <Rle>
#>    [1]        I 151583-151641      +
#>    [2]       II 238361-238419      +
#>    [3]      III 114322-114380      +
#>    [4]       IV 449879-449937      +
#>    [5]        V 152522-152580      +
#>    ...      ...           ...    ...
#>   [12]      XII 151366-151424      +
#>   [13]     XIII 268222-268280      +
#>   [14]      XIV 628588-628646      +
#>   [15]       XV 326897-326955      +
#>   [16]      XVI 556255-556313      +
#>   -------
#>   seqinfo: 17 sequences (1 circular) from R64-1-1 genome
pairsFile(full_contacts_yeast)
#> NULL
anchors(contacts_yeast)
#> $first
#> GRanges object with 74360 ranges and 2 metadata columns:
#>           seqnames        ranges strand |   chr    center
#>              <Rle>     <IRanges>  <Rle> | <Rle> <integer>
#>       [1]       II        1-1000      * |    II       500
#>       [2]       II        1-1000      * |    II       500
#>       [3]       II        1-1000      * |    II       500
#>       [4]       II        1-1000      * |    II       500
#>       [5]       II        1-1000      * |    II       500
#>       ...      ...           ...    ... .   ...       ...
#>   [74356]       II 807001-808000      * |    II    807500
#>   [74357]       II 807001-808000      * |    II    807500
#>   [74358]       II 808001-809000      * |    II    808500
#>   [74359]       II 808001-809000      * |    II    808500
#>   [74360]       II 809001-810000      * |    II    809500
#>   -------
#>   seqinfo: 16 sequences from an unspecified genome
#> 
#> $second
#> GRanges object with 74360 ranges and 2 metadata columns:
#>           seqnames        ranges strand |   chr    center
#>              <Rle>     <IRanges>  <Rle> | <Rle> <integer>
#>       [1]       II     1001-2000      * |    II      1500
#>       [2]       II     5001-6000      * |    II      5500
#>       [3]       II     6001-7000      * |    II      6500
#>       [4]       II     8001-9000      * |    II      8500
#>       [5]       II    9001-10000      * |    II      9500
#>       ...      ...           ...    ... .   ...       ...
#>   [74356]       II 809001-810000      * |    II    809500
#>   [74357]       II 810001-811000      * |    II    810500
#>   [74358]       II 808001-809000      * |    II    808500
#>   [74359]       II 809001-810000      * |    II    809500
#>   [74360]       II 809001-810000      * |    II    809500
#>   -------
#>   seqinfo: 16 sequences from an unspecified genome
#> 
regions(contacts_yeast)
#> GRanges object with 802 ranges and 2 metadata columns:
#>                    seqnames        ranges strand |   chr    center
#>                       <Rle>     <IRanges>  <Rle> | <Rle> <integer>
#>          II_1_1000       II        1-1000      * |    II       500
#>       II_1001_2000       II     1001-2000      * |    II      1500
#>       II_5001_6000       II     5001-6000      * |    II      5500
#>       II_6001_7000       II     6001-7000      * |    II      6500
#>       II_7001_8000       II     7001-8000      * |    II      7500
#>                ...      ...           ...    ... .   ...       ...
#>   II_807001_808000       II 807001-808000      * |    II    807500
#>   II_808001_809000       II 808001-809000      * |    II    808500
#>   II_809001_810000       II 809001-810000      * |    II    809500
#>   II_810001_811000       II 810001-811000      * |    II    810500
#>   II_812001_813000       II 812001-813000      * |    II    812500
#>   -------
#>   seqinfo: 16 sequences from an unspecified genome
summary(contacts_yeast)
#> sparse `contacts` object with 74,360 interactions over 802 regions 
show(contacts_yeast)
#> sparse `contacts` object with 74,360 interactions over 802 regions 
#> coolPath: /home/rsg/.cache/R/ExperimentHub/36d590c5583_7752 
#> focus: II 
#> ------------
#> metadata(0):
#> resolutions(5): 1000 2000 4000 8000 16000
#> current resolution(1000): 74360 interactions in memory 
#> bins: 12079 
#> regions: 802 
#> scores(3): raw balanced test 
#> features: loops(0) borders(0) compartments(0) viewpoints(0) centromeres(16) 
#> pairs: N/A